@@ -1,11 +1,8 @@
 ﻿@page "/"
 
-@using System.Net.Http
-@inject HttpClient Http
-@using System.Net.Http.Json
-@using System.Threading.Tasks
-@using DocumentApp.DTO
-
 <PageTitle>Список публикаций</PageTitle>
+<PageHeader>Список публикаций</PageHeader>
 
-<PublicationListComponent />
+<PublicationsPresenter />
+
+<hr/>
