@@ -1,16 +1,9 @@
 ﻿@page "/edit"
 
-@using System.Net.Http
-@inject HttpClient Http
-@using System.Net.Http.Json
-@using System.Threading.Tasks
-@using DocumentApp.DTO
-@inject NavigationManager _nav
-
 <PageTitle>Редактирование публикации</PageTitle>
 
-<div class="Block_C">
-    <h2 class="list__source__title">Редактирование публикации</h2>
+<div class="page-header-container">
+    <h2 class="page-header">Редактирование публикации</h2>
 </div>
 
 <PublicationsPresenter />
